+++ test/shopify-cli/theme/dev_server_test.rb
@@ -118,9 +118,10 @@ module ShopifyCLI
       private
 
       def dev_server(identifier: nil, ignores: nil, includes: nil)
-        host, port, poll, editor_sync, stable, mode = nil
+        host, port, poll, editor_sync, overwrite_json, stable, mode = nil
         server = DevServer.instance
-        server.setup(ctx, root, host, identifier, port, poll, editor_sync, stable, mode, includes, ignores)
+        server.setup(ctx, root, host, identifier, port, poll, editor_sync, overwrite_json, stable, mode, includes,
+          ignores)
         server
       end